Current working state #noissue
I've created a new talk detail page that includes the event slug in the
URL.  I've added some ideas to the README how the URLs should be
structured. A lot of work still has to be done to make the attendee and
speaker registration work correctly, and I think there's still a number
of views that need to be implemented.

I did not manage to write any tests.