SECO-56 Scripts for production Vault setup
This commit adds scripts for production Vault setups.

init_unseal_fill_Vault.sh is a script to properly initialize a Vault
instance for the devday application. It performs the following steps:

- initialize Vault with 3 key shares and a threshold of 2 required
  shares
- store key shares and root_token in a json file
- unseal Vault with the two first keys from the json file
- enable Vault audit logging to /Vault/logs/audit.log
- make sure that the default secret backend uses the kv version 2
  versioning support
- define a Vault policy to properly restrict the application user to
  /secret/data/devday
- put initial data in /secret/data/devday
  - a newly generated postgresql_password
  - a newly generated secret_key
- create a new application token and store it into prod-env to be used
  for the app container by docker-compose

The script takes care of not overwriting any data. It can be run again
after initialization of Vault to:

- unseal Vault after it has been restart
- upload new policy versions if changes are needed
- create new devday entry if it has been deleted
- create new application token if the old one has been
  lost/revoked/expired

devday_policy.hcl contains the Vault policy for the devday application
user.

renew_app_token.sh can be used to renew the application user token
before it expires. This script should be run by a cron job to avoid the
expiration of the application token.

Note:

  A new application token has to be created if the application token has
  expired. A root token is required to create a new application.
  init_unseal_fill_vault.sh automates the creation of a new application
  token if the generated json file from the first run of that script is
  available.