Add attendee checkin
For the attendee, a new view `/attendee/qrcode/` shows a QR code that
the attendee can show at the event to be scanned to check in. The same
information has been added to the user profile page.

Staff at the event (users with `is_staff=True`) can scan the QR code
with their phone and will automatically check the attendee in. A page
will show if the attendee was checked in successfully, or an indication
why the checkin failed.

In case the user did not bring the QR code, a manual checkin page
`/attendee/checkin/` can be used to check in the attendee using either
their email address, or an 8-digit code unique to that attendee, to make
entering the details easier. The code is shown on the user's profile
page and on the QR code page. This page is shown in the menu as
"Check In" if the user is staff.