Unify prod and dev Dockerfile for fast image builds
This commit unifies the Dockerfile-dev and Dockerfile-prod so that both
use as much common lines as possible. This improves image build speed.

The COPY command for the application code in the prod Dockerfile has
been moved to the end of the file so that rebuilds of earlier layers are
only necessary if dependencies in Pipfile/Pipfile.lock or the base image
change.

The dev settings have been changed so that MEDIA_ROOT and STATIC_ROOT
are mapped to a higher directory level. They now land in
/srv/devday/media and /srv/devday/static and are therefore the same in
dev and prod application containers. The mounts for these directories in
docker-compose.dev.yml have been adapted accordingly.