Separate non-attendee registration flow
This commit introduces a separate registration flow for non-attendees.

The new view attendee.views.DevDayUserRegistrationView accepts a URL
parameter "next" that is passed on to the registration form as well as
the activation email sent to the new user.

The setting SECURE_PROXY_SSL_HEADER has been moved to
devday.settings.base to allow tests with https scheme URLs.