Address review comments
This commit adds a editable static placeholder to explain the sponsoring
oportunities in more details. The contact form provides a choice of sponsoring
options and adds the chosen options to the email that is sent to the sponsoring
team.

Tests and translations have been updated to match the new implementation.