Handle closed attendee registration
This commit adds a new AttendeeRequiredMixin and changes existing views
in the attendee and talk apps to use this mixin. The mixin makes sure
that views that require an attendee redirect to the attendee
registration view as long as the event allows attendee registration. If
the attendee registration is closed the mixin redirects to the
django_registration_disallowed url.

The commit also ensures that knowing the attendee registration URL is
not sufficient to register an attendee and that no attendee activation
is possible when the attendee registration has been closed.

All tests have been adapted, new tests have been added.

This fixes #160