Allow 0.0.0.0 to Access Page in Development
The Docker container starts the Django application by binding
it on 0.0.0.0. This is required to allow accessing the application
from "outside" of the container (aka the developers computer).

The Docker container will print a message that tells the
developer that the app is available at "http://0.0.0.0:8000".

Unfortunately, this is "IP" is not a "allowed host" out-of-the-box.
Therefore, this "IP" has to be added to the "ALLOWED_HOSTS" array
in the configuration.

With this change you can click on the link that Djangos
"./manage.py runserver 0.0.0.0:8000" prints and access the page.