Restructure docker images
Alpine switched the default Python to Python 3.8 which broke Django 2.1
and our own docker images. Images have now been rebuilt to use a common
base image based on Alpine instead of python:alpine.

I had to use a fork of djangocms-column that works with Django 2.2.

I moved the application Dockerfiles to the top level directory. And
modified run.sh and the docker-compose files that are used for
development builds.