Use Slug for Session URL
Minor update to the existing code for the session list. The only real
change is that the "id" is now the slug of the Talk and the "url" uses
the talks slug instead of the pk.

Further improvements will be made to the serializer, once the other
endpoints/serializers are done.