Add Initial Speakers Endpoint
The endpoint outputs basic information about speakers, including an
endpoint for speaker details. At the moment there is no relationship to
other entities yet. This will be added at a later point in time, like
for example the talks the speaker has given or will give or the events
the speaker presented at.