Add 'events' API Endpoint
The is a draft of a possible Events API. Not all fields of the Event
model are exposed. This is by design, to not over-expose too much
information, that nobody needs.

This will be developed over time. One good extension might be to add
links to all the sessions associated with the event.

Furthermore, the API could be extended by "actions". Links that can be
followed to for example register for an event etc.