Add Speaker Information to Events API
Instead of just containing a list of speaker names, after this patch is
applied, the API will contain a list of objects. Object object per
(published) speaker. The object will contain a URL to the speakers
details (in the API), the name of the speaker and image URL of the
speakers thumbnail.