Introduce Swagger UI and ReDoc API Documentation
This commit adds Swagger UI endpoint at 'api/schema/swagger-ui/'. To
achieve this a new dependency called 'drf-spectacular' was added.

Currently, the documentation requires authentication by the user, but
this might be changed in the future. This was done because the API
itself requires authentication.

The configuration for drf-spectacular is mostly set to default, only
authentication and schema exposure have been changed. Those changes have
been added to the end of 'devday/settings/__init__.py'.

See: https://drf-spectacular.readthedocs.io