Enable Token Authentication for API
Minor change to enable Token authentication for our API. This is meant
to enable applications that do not want to use the typical
username/password authentication scheme using an underlying session.

Our SPA for example would use the tokens to be able to only communicate
with the API and do authenticated requests in the name of the user.